Send new users to add course
We want users to add a course straight away, or fail fast – ie leave for UCAS if we can't accommodate their choice.

We don't want to risk users completing an application form, then not being able to apply to the course they want to.

Introduces concept of a welcome flow.