Change "Conditions met" to "Offer confirmed"
Currently we display "Conditions met" as the status for courses where the provider has confirmed that the candidate has met all the conditions required for the course.

However, some courses don’t have any conditions at all, or at least any conditions specified on the service (there might be some administrative conditions off-service). In these scenarios, the status doesn’t make sense.

Whilst we _could_ have a different status label for those who have accepted unconditional offers only, it would be simpler if we could find a new label that works in all scenarios.

"Offer confirmed" is potentially a more positive-sounding label than "Conditions met", and is also consistent with the language used on the service when the candidate has not yet met the conditions, which states:

> You’ve accepted the offer from [provider name]
>
> Before your offer is confirmed, there are some final conditions to meet.