Remove "Do you want to add another course?" page
...instead, redirect users to the course choices page, from where they can add another course if they are able to do so.

This page was added to encourage candidates to apply to multiple courses.  However since then we have moved the "Add another course" secondary button to the top of the course choices page, and added some hint text to the "Have you completed this section?" radio button.

This also adds the same guidance text to just above the "Add another course" button.

Our hypothesis is that this will serve the same purpose as the "Do you want to add another course" page, whilst removing an unnecessary extra page and giving users extra reassurance they just added the right course.

An open question is whether we still need the "Success" banner or not – with an initial hypothesis being that we don't, as the page itself is a confirmation.