Make it clear you can return to reference (#548)
We currently tell referees that they need to complete the reference in a "one sitting". However they can save and come back to it at any time, using the original link sent to their email.

Feedback from referees has indicated that they find the requirement to complete the reference in a single sitting to be stressful, especially given that the reference takes longer than they initially expect. Some referees may work around this by copying and pasting to and from a separate file, but this can be cumbersome, and means they might not see the word count straight away.

To address this, we are changing the content to make clear that they can return to the form after saving it. The buttons are also relabelled from "Continue" to "Save and continue". However the button on the page with the 500 word textarea is just labelled "Save" so that it's clear you can save this even if you aren't ready to continue.

The check your answers page also has new content to make it clearer that referees can exit at this page and return to continue, rather than submit it now, and the title changed from "Check your answers before submitting your reference" to "Your reference for [candidate name]".

The title of the reference page has changed from "Your reference for [candidate name]" to a question: "Does Jane Doe have the potential to teach?". This is to distinguish it from the last page, and because elsewhere we use "reference" to refer to the whole journey, not just this question.

The bullet points on the reference page have also been revised to be more succinct, and more in keeping with the style guide. (Previously the bullets contained headings and paragraphs).