Simplify other qualifications flow
This simplifies the "What type of qualification do you want to add?" page to just have the question "What type of qualification is it?" with no option of "I do not want to add any A levels and other qualifications".

Instead, candidates are routed through the review page, which contains the guidance about why it is a good idea to add A levels and other qualifications.