Add fee information to offer details (#603)
This would initially just copy across the content from the Fees or Salary section on Find.  The amount of content can vary quite a bit between providers.

This information is required as part of the pre-contract information by the regulations. We probably fulfil that already through the link to the course page on Find, plus any additional information given direct to candidates by providers off-service. 

However, it still feels like a good idea to repeat this on the offer page, to remind candidates of the fees / salary at the point of accepting an offer.