Update ethnicity questions
This updates the ethnicity questions to bring them in line with the [GOV.UK Design System pattern for equality information](https://design-system.service.gov.uk/patterns/equality-information/) - with the exception that Roma has been added as an option, following the addition of that option to the 2021 census and the HESA ITT ethnic group categories.

The other changes are ordering changes, and changing "Another" to "Any other" and "Your Black background (optional)" to "How would you describe your background? (optional)".