Show message if you can't apply to a course (#726)
This would allow candidates to add courses which aren't yet open for applications, but would warn them and prevent the application being submitted.

This would hopefully make it more obvious (than just not allowing them to find the course in the dropdowns/radios), and allows candidates to get their whole application ready, waiting for the course to open. Alternatively they can choose a different course.

We can use the same approach for when a course is "full" / has no vacancies.