Move equality and diversity questions into application (#784)
This moves the equality and diversity questions from being answered just before submitting to being part of the main application task list. This is in preparation for continuous applications, where each course choice would be submitted separately, and so we wouldn't want to re-ask the same equality and diversity questions each time.