FAC opt in updates (#913)
* Find a candidate updates

Problem:
Nobody is reading the explanatory content on the pre-opt-in page
Additional content trying to explain the details made people less likely to read/understand the core proposition
It's unclear when they are in and out of the list
There's an impression that they can submit more than 4 applications and that they'll have a higher chance of success/be prioritised in the process because of candidate pool

Suggested solution:
Remove the pre-opt-in page. Supply essential information on the opt-in page and detailed information afterwards.
Explain when candidates are in/out of the list in the detailed information.
Update the section on Your applications to show when someone is opted in/out and visible/not visible to explain this concept in context
Change the H1 on the opt-in to remove reference to 'increasing chances of success' and make it clearer what you are opting in to (being invited to similar courses to the one you've just applied to)

Problem:
International candidates are setting high radii instead of no location preferences to cover being open to courses anywhere
Participants were not clear about what the search radius meant in reality
The checkbox to add new course locations to preferences was not well understood by everyone

Suggested solution:
Add a question about training locations at the start of the process to capture everyone who has no preferences
Change 'course locations' to 'areas you can train in'
Change the table headers and content around search radius to reference travel time
Move add course locations to my preferences to its own page

* FAC updates

* Fixing back links